feat: validations exist yarn and pm2 file build

+++ build.sh
@@ -1,38 +1,44 @@
 #!/bin/sh
 
-echo "REMOVING dist..."
-rm -rf dist/
-echo "dist REMOVED!"
-
-echo "--------------------------------"
-
-echo "INSTALLING DEPENDECIES..."
-yarn
-echo "DEPENDECIES INSTALLED!"
-
-echo "--------------------------------"
-
-echo "BUILD APLICATION..."
-yarn build
-echo "BUILD SUCCESS!"
-
-echo "--------------------------------"
-
-if [ ! -d "dist" ]; then
-  echo "CRIANDO PASTA DIST"
-  mkdir dist
+if ! type "yarn" > /dev/null; then
+  echo "\e[41m INSTALL yarn  \e[0m"
+  exit
 fi
 
-VERSION=$(jq -r '.version' package.json)
-DATE=$(date -u +"%Y-%m-%dT%H:%M:%SZ")
-jq -n --arg version $VERSION --arg dateValue $DATE '{"version" : $version , "created_at" : $dateValue}' > dist/version.json
+echo "--------------------------------"
+echo "\e[33m REMOVING dist... \e[0m"
+rm -rf dist/
+echo "\e[32m dist REMOVED! \e[0m"
 
 echo "--------------------------------"
 
-echo "RESTARTING SERVER..."
-pm2 restart notifier-container-docker
-echo "SERVER RESTARTED!"
+echo "\e[33m INSTALLING DEPENDECIES... \e[0m"
+yarn
+echo "\e[32m DEPENDECIES INSTALLED! \e[0m"
 
 echo "--------------------------------"
 
-echo "BUILD SUCCESS!!!"
+echo "\e[33m BUILD APLICATION... \e[0m"
+yarn build
+echo "\e[32m BUILD SUCCESS! \e[0m"
+
+echo "--------------------------------"
+if type "pm2" > /dev/null; then
+  if [ ! -d "dist" ]; then
+  echo "\e[33m CREATING DIST PASTE... \e[0m"
+    mkdir dist
+  fi
+
+  VERSION=$(jq -r '.version' package.json)
+  DATE=$(date -u +"%Y-%m-%dT%H:%M:%SZ")
+  jq -n --arg version $VERSION --arg dateValue $DATE '{"version" : $version , "created_at" : $dateValue}' > dist/version.json
+
+  echo "--------------------------------"
+  echo "\e[33m RESTARTING SERVER... \e[0m"
+  pm2 restart notifier-container-docker
+  echo "\e[32m SERVER RESTARTED! \e[0m"
+
+  echo "--------------------------------"
+fi
+
+echo "\e[32m --> BUILD FINISH <-- \e[0m"
